FC Modularity
Since the friendcode command essentially grabs some value from a file, I rolled a promo command into it's functionality in the process of making it entirely modular. Does it not make much sense to say Promo switch to get the switch friendcode? yeah, but nobody's going to use it like that, they'll use the command that makes sense.
