+++ src/main/java/com/glacier/bot/Robot.java
@@ -2,8 +2,8 @@ package com.glacier.bot;
 
 import java.io.File;
 
+import com.glacier.commands.CodesAndPromos;
 import com.glacier.commands.DiscordAnnouncement;
-import com.glacier.commands.FriendCodes;
 import com.glacier.commands.Gutenberg;
 import com.glacier.commands.Ping;
 import com.glacier.commands.Uptime;
@@ -40,7 +40,7 @@ public class Robot {
 		twitchClient.getCommandHandler().registerCommand(Gutenberg.class);
 		twitchClient.getCommandHandler().registerCommand(clearMessage.class);
 		twitchClient.getCommandHandler().registerCommand(DiscordAnnouncement.class);
-		twitchClient.getCommandHandler().registerCommand(FriendCodes.class);
+		twitchClient.getCommandHandler().registerCommand(CodesAndPromos.class);
 		//the setstart command really shouldnt have to exist but I can't find a way to track uptime in this api
 		//so instead I'm having the start time manually get set by changing the last updated at
 		//besides it's best to have a manual start time anyway since then I can cut out the "setup" time
